requirements.txt
pycryptodome==3.15.0
sympy==1.13.2
gmpy2==2.2.1